♻️ Refactor: 유틸함수로 수정
