+++ src/pages/HouseWorkStepOnePage.tsx
@@ -100,7 +100,7 @@ const HouseWorkStepOnePage = () => {
             text='어떤 집안일인가요?'
             handleClick={handleHouseWorkClick}
             type='housework'
-            icon={<EtcIcon fillOneClass='fill-gray1' fillTwoClass='fill-gray3' />}
+            icon={<EtcIcon fillOneClass='fill-gray2' fillTwoClass='fill-gray3' />}
           />
         )}
         {startDate ? (
